dashboard/backend/requirements.txt
# Core Django & Server
Django==5.0.3
gunicorn==21.2.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
whitenoise==6.6.0

# Database
pymongo==4.6.2
dnspython==2.6.1  # Required for Atlas srv:// connection

# Data Processing (Kept from your list)
pandas==2.2.1
numpy==1.26.4
requests==2.31.0
pyOpenSSL==24.0.0
cryptography==42.0.5
reportlab==4.1.0
pillow==10.2.0

# Utilities
pytz==2024.1
urllib3==2.2.1
certifi==2024.2.2